Address Jonathan's comments
